+++ packages/verified-fetch/src/utils/parse-resource.ts
@@ -1,7 +1,7 @@
 import { CID } from 'multiformats/cid'
 import { parseUrlString } from './parse-url-string.js'
 import type { ParsedUrlStringResults } from './parse-url-string.js'
-import type { ResourceType } from '../interface.js'
+import type { ResourceType } from '../index.js'
 import type { IPNS } from '@helia/ipns'
 /**
  * Handles the different use cases for the `resource` argument.
